pyproject.toml
[tool.poetry]
name = "isezakiiroha"
version = "0.0.1"
description = "Isezaki Ikuho"
authors = ["nothink <gakongakon@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
djangorestframework = "^3.11.1"
django-environ = "^0.4.5"
google-cloud-storage = "^1.31.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = { version = "*", allow-prereleases = true }
isort = "^5.4.2"
flake8-black = "^0.2.1"
pytest = "^6.0.1"
pytest-django = "^3.9.0"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
